requirements.txt
# ============================================================
# FaultTreeAI - 外部API + LangChain + PostgreSQL 版本
# ============================================================

# Web框架
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# LangChain 核心
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# PostgreSQL + pgvector
psycopg2-binary>=2.9.10
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.30
pgvector>=0.3.0
langchain-postgres>=0.0.17

# MiniMax API（项目自行封装，不依赖第三方 langchain-minimax）
# MiniMax SDK（HTTP 调用封装）
aiohttp>=3.9.0

# 文档解析
PyMuPDF>=1.24.0
python-docx>=1.1.0

# 图计算
networkx>=3.3
numpy>=1.26.0

# HTTP客户端
httpx>=0.27.0

# 报告导出
reportlab>=4.2.0

# 工具
python-multipart>=0.0.9
aiofiles>=24.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
tenacity>=9.0.0
tiktoken>=0.12.0

# 关键词检索
rank-bm25>=0.2.2
